pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["Bekbolot001 <bekbolotboronchiev.bt.966@gmail.com>"]
readme = "README.md"
packages = [{include = "blog_api"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.3.0"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
django-jazzmin = "^3.0.0"
celery = "^5.4.0"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
